requirements.txt
asgiref==3.7.2
attrs==23.2.0
black==24.1.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coreschema==0.0.4
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-redis==5.4.0
django-soft-delete==1.0.12
djangorestframework==3.14.0
djangorestframework-api-key==3.0.0
djangorestframework-gis==1.0
djangorestframework-simplejwt==5.3.1
drf-extensions==0.7.1
drf-spectacular==0.27.1
flake8==6.1.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.31.1
requests==2.31.0
rpds-py==0.17.1
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.2.1
